requirements.txt
google-genai>=1.0.0
pymupdf>=1.24.0
firebase-admin>=6.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm
